+++ src/composition/debounce.js
@@ -1,0 +1,9 @@
+import { ref } from "vue";
+
+import Debounce from "../tools/Debounce";
+
+export default function useDebounce() {
+  const DebounceConstructor = ref(Debounce);
+
+  return { DebounceConstructor };
+}