+++ src/composition/emitter.js
@@ -1,9 +1,9 @@
 import { ref } from "vue";
 
-import LocalEmitter from "../events/LocalEmitter";
+import LocalEmitter from "../tools/LocalEmitter";
 
 export default function useEmitter() {
-    const emitter = ref(new LocalEmitter());
+  const emitter = ref(new LocalEmitter());
 
-    return { emitter };
+  return { emitter };
 }