+++ src/events/instance.js
@@ -1,3 +1,0 @@
-import LocalEmitter from "../events/LocalEmitter";
-
-export default new LocalEmitter();