+++ src/tools/instance.js
@@ -1,0 +1,3 @@
+import LocalEmitter from "../tools/LocalEmitter";
+
+export default new LocalEmitter();